Moved RecordReaderGrakn into InputFormatGrakn
